refactor: elderly dto 리팩토링
