--- src/main/java/com/becareful/becarefulserver/domain/matching/dto/response/InstitutionInfoResponse.java
@@ -1,10 +1,0 @@
-package com.becareful.becarefulserver.domain.matching.dto.response;
-
-import com.becareful.becarefulserver.domain.nursingInstitution.domain.NursingInstitution;
-
-public record InstitutionInfoResponse(String name, String address) {
-    public static InstitutionInfoResponse from(NursingInstitution institution) {
-        return new InstitutionInfoResponse(
-                institution.getName(), institution.getAddress().getFullAddress());
-    }
-}
