+++ src/main/java/com/becareful/becarefulserver/domain/matching/dto/response/MyRecruitmentDetailResponse.java
@@ -1,6 +1,8 @@
 package com.becareful.becarefulserver.domain.matching.dto.response;
 
 import com.becareful.becarefulserver.domain.matching.domain.Recruitment;
+import com.becareful.becarefulserver.domain.matching.dto.ElderlyDto;
+import com.becareful.becarefulserver.domain.matching.dto.InstitutionDto;
 import com.becareful.becarefulserver.domain.nursingInstitution.domain.NursingInstitution;
 import com.becareful.becarefulserver.domain.socialworker.domain.Elderly;
 import java.time.LocalDate;
@@ -20,8 +22,8 @@ public record MyRecruitmentDetailResponse(RecruitmentDetailResponse recruitmentD
         return new MyRecruitmentDetailResponse(
                 new RecruitmentDetailResponse(
                         RecruitmentInfoResponse.from(recruitment, elderly),
-                        ElderlyInfoResponse.from(elderly),
-                        InstitutionInfoResponse.from(institution),
+                        ElderlyDto.from(elderly),
+                        InstitutionDto.from(institution),
                         isHotRecruitment,
                         isHourlySalaryTop,
                         matchRate),
