+++ src/main/java/com/becareful/becarefulserver/domain/matching/dto/response/RecruitmentDetailResponse.java
@@ -1,13 +1,15 @@
 package com.becareful.becarefulserver.domain.matching.dto.response;
 
 import com.becareful.becarefulserver.domain.matching.domain.Recruitment;
+import com.becareful.becarefulserver.domain.matching.dto.ElderlyDto;
+import com.becareful.becarefulserver.domain.matching.dto.InstitutionDto;
 import com.becareful.becarefulserver.domain.nursingInstitution.domain.NursingInstitution;
 import com.becareful.becarefulserver.domain.socialworker.domain.Elderly;
 
 public record RecruitmentDetailResponse(
         RecruitmentInfoResponse recruitmentInfo,
-        ElderlyInfoResponse elderlyInfo,
-        InstitutionInfoResponse institutionInfo,
+        ElderlyDto elderlyInfo,
+        InstitutionDto institutionInfo,
         boolean isHotRecruitment,
         boolean isHourlySalaryTop,
         Integer matchRate) {
@@ -19,8 +21,8 @@ public record RecruitmentDetailResponse(
 
         return new RecruitmentDetailResponse(
                 RecruitmentInfoResponse.from(recruitment, elderly),
-                ElderlyInfoResponse.from(elderly),
-                InstitutionInfoResponse.from(institution),
+                ElderlyDto.from(elderly),
+                InstitutionDto.from(institution),
                 isHotRecruitment,
                 isHourlySalaryTop,
                 matchRate);
